Unique IDs
Not quite clear what was going on in my little pea brain… but _now_ there are unique IDs for each entry in `app_metadata` that are _not_ the actual `key` values -- which allows one to have multiple apps in the same database (a key design feature) with identical keys _per app_ without collisions. Duh.

Also, ditched the schema file in favor of the software-driven update (which allows the name of the `app_metadata` table to be configurable, as it should have been (but was not). Also provides the capability to update older databases when the software is updated… and loses the static call to `prepareDatabase()` (which means that a truly honest semantic versioner would have made this 2.0… but I'm a dishonest lout, so I'm  just going with 1.2, since it's not a _big_ upgrade and _I'm_ the only one using this right now. I think.)